Restructure: extract Query typestate + RangeJourney into query.rs
